requirements.txt
ariadne==0.20.1
fastapi==0.96.0
boto3==1.28.59
botocore==1.31.59
requests==2.31.0
backoff==2.2.1
python-dotenv==1.0.0
cpeparser==0.0.2
uvicorn==0.17.6
npm==0.1.1
click==8.1.7
minio==7.2.9
pyarrow==17.0.0
dpath==2.2.0
pyyaml==6.0.2
awswrangler==3.9.1
aws-secretsmanager-caching